requirements.txt
Pillow
tkinter
ttkbootstrap
requests>=2.31.0
pywin32>=300; platform_system == "Windows"
python-dotenv>=0.19.0